requirements.txt
anyio==3.5.0
asgiref==3.5.0
autopep8==1.6.0
click==8.0.4
ecdsa==0.17.0
fastapi==0.75.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.2
install==1.3.5
passlib==1.7.4
psycopg2==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
pydantic==1.9.0
python-dotenv==0.19.2
python-jose==3.3.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
spark-parser==1.8.9
SQLAlchemy==1.4.4
starlette==0.17.1
toml==0.10.2
typing_extensions==4.1.1
uncompyle6==3.8.0
unpyclib==0.8.1
uvicorn==0.17.5
xdis==6.0.3
zipp==3.7.0
